Roomba Coding Challenge
There is a room with 8 by 8 tiles and the roomba is placed in the south-western corner. There are 18 random dirty tiles. The roomba has to move over the tile to clean it.Each move the roomba makes it loses 1/1000 of its battery. IF the battery is empty the roombas task will be unsucsessfull. Every clean block the roomba moves  over 10 marks are deducted and every tile it cleans 250 marks are added.